initial commit
Here is a Solidity interface that allows you to access information from the PLPBondDistributor and PLPToken contracts:

Copy code

The PLPInfoInterface contract defines the functions that are common to both the PLPBondDistributor and PLPToken contracts, such as multisig, treasury, totalPLP, totalLP, and bondedPLP. It also defines the balanceOf function, which is part of the ERC20 standard and allows you to check the balance of a specified address.

The PLPBondDistributor interface extends the PLPInfoInterface and defines the additional functions specific to the PLPBondDistributor contract, namely swapPLPForLP and claimLP.

The PLPToken interface extends the SafeERC20 interface (provided by OpenZeppelin) and the PLPInfoInterface, and defines the additional mint function specific to the PLPToken contract.

You can use these interfaces to access the functions of the PLPBondDistributor and PLPToken contracts in your own contracts or scripts. For example, you could use the following code to call the swapPLPForLP function of the PLPBondDistributor contract:

Copy code
PLPBondDistributor plpBondDistributor = PLPBondDistributor(plpBondDistributorAddress);
plpBondDistributor.swapPLPForLP(100);

I hope this helps! Let me know if you have any questions or if you'd like further clarification.