initial commit
Here is a sample hardhat.config.js file for this project:

Copy code

This configuration file sets the default network to "localhost" and defines a local blockchain network with a hardcoded mnemonic. It also specifies the use of the hardhat-deploy plugin to deploy the PLPToken and PLPBondDistributor contracts to the local blockchain.

The constructorArgs property specifies the arguments to pass to the contract's constructor function when it is deployed. In this case, the PLPToken contract is passed the name, symbol, decimals, and total supply arguments, and the PLPBondDistributor contract is passed the multisig and treasury address arguments.

You will need to replace the your mnemonic here placeholder with your own mnemonic, and specify the addresses of the multisig and treasury contracts in the constructorArgs of the PLPBondDistributor contract.

I hope this helps! Let me know if you have any questions or if you'd like further clarification.