initial commit
R: OK one last file : please define an index.html file for our server.js to connect to and display to users.

The index.html needs to be able to interact *only* with public functions in plpBondDistributor.sol . It should also have a connect button for users' browser wallet extensions, and display the following information to users:

- current user balance of PLP
- current user balance of LP token
- current Total Value Locked (TVL)
- current PLP token price
- current LP token price
- current ratio of PLP:LP token prices (PLP price divided by LP token price)
- current ratio of PLP:TotalTreasury (PLP price divided by Total Treasury)
- explain the 1:0.9 fixed swap rate ("This allows our protocol to make a flat 10% profit on every swap and therefore stay in business.")
- display the total amount (in USD) in the treasury address
- display the total amount (in USD) in the multisig address

thank you.

chatgpt: Here is a sample index.html file that you can use to display the information and interact with the plpBondDistributor contract:

Copy code

This HTML file includes placeholders for the various pieces of information that you requested (e.g., user balances, token prices, TVL, etc.), as well as buttons for connecting a wallet and interacting with the plpBondDistributor contract. The index.js script file, which you will need to create separately, will be responsible for fetching the data and updating the DOM elements with the correct values.