initial commit
This code looks correct! Here are a few things to note:

The plpToken and plpBondDistributor contract instances are connected to the user's wallet provider when the user clicks the "connect" button. This allows the user to interact with the contracts using their own wallet.

The bondLPButton and bondPLPButton buttons are disabled by default, and are only enabled when the user clicks the "connect" button. This ensures that the user has a connected wallet before they can interact with the contract.

The bondLPButton and bondPLPButton buttons have event listeners attached to them, which call the bondLP and bondPLP functions respectively when clicked.
The getUSDValue function is not defined in this code, so you will need to make sure that it is defined elsewhere in your code. This function is used to convert the value of the totalTreasury and multisig variables to USD.

Overall, it looks like this frontend/index.js file is set up to interact with the plpBondDistributor contract and update the DOM elements in the index.html file with the relevant data. You may want to add additional functionality or error handling as needed for your application.